feat: 채팅방 목록 수정
