+++ client/src/styles/globals.css
@@ -24,6 +24,18 @@
         background-color: rgb(0,0,0,0.2);
         border-radius: 100px;
     }
+
+    .custom-scroll-bar-6px::-webkit-scrollbar-thumb:hover {
+        background-color: rgb(0,0,0,0.3);
+    }
+    .custom-scroll-bar-6px::-webkit-scrollbar {
+        width: 6px;
+        height: 6px;
+    }
+    .custom-scroll-bar-6px::-webkit-scrollbar-thumb {
+        background-color: rgb(0,0,0,0.2);
+        border-radius: 100px;
+    }
     .custom-scroll-bar-10px::-webkit-scrollbar-thumb:hover {
         background-color: rgb(0,0,0,0.3);
     }
